List the traits available to add_trait()

Returns the traits that add_trait() can attach across sources, with their kind, canonical unit, and the number and names of contributing sources.

Usage

list_traits()

Value

A data.frame with one row per trait:

trait

The trait name to pass to add_trait().

label

Human-readable label.

kind

"numeric" or "categorical".

unit

Canonical unit for numeric traits, NA for categorical.

n_sources

Number of sources providing the trait.

sources

Comma-separated source names.

See also

Examples

list_traits()
#>                             trait                                    label
#> 1                    plant_height                             Plant height
#> 2                       seed_mass                                Seed mass
#> 3                             sla                       Specific leaf area
#> 4                    wood_density                             Wood density
#> 5                       leaf_area                                Leaf area
#> 6                     leaf_length                              Leaf length
#> 7                      leaf_width                               Leaf width
#> 8                          leaf_n               Leaf nitrogen per dry mass
#> 9                          leaf_p             Leaf phosphorus per dry mass
#> 10                 leaf_thickness                           Leaf thickness
#> 11                  leaf_lifespan                            Leaf lifespan
#> 12                 plant_lifespan                     Whole-plant lifespan
#> 13                      body_mass                                Body mass
#> 14                      longevity                        Maximum longevity
#> 15                  trophic_level                            Trophic level
#> 16             clutch_litter_size                    Clutch or litter size
#> 17                age_at_maturity                   Age at female maturity
#> 18                  male_maturity                     Age at male maturity
#> 19           gestation_incubation           Gestation or incubation length
#> 20              incubation_period                    Egg incubation period
#> 21                    body_length                              Body length
#> 22                 metabolic_rate                           Metabolic rate
#> 23                     brain_mass                               Brain mass
#> 24         reproductive_frequency             Litters or clutches per year
#> 25            interbirth_interval      Interbirth or inter-litter interval
#> 26                    teat_number                    Teat or nipple number
#> 27             population_density                       Population density
#> 28                   neonate_mass                        Neonate body mass
#> 29                       egg_mass                                 Egg mass
#> 30                     egg_length                               Egg length
#> 31                      egg_width                                Egg width
#> 32                flowering_start                  Flowering start (month)
#> 33                  flowering_end                    Flowering end (month)
#> 34                ellenberg_light                      Ellenberg light (L)
#> 35          ellenberg_temperature                Ellenberg temperature (T)
#> 36             ellenberg_moisture                   Ellenberg moisture (F)
#> 37             ellenberg_reaction                   Ellenberg reaction (R)
#> 38             ellenberg_nitrogen       Ellenberg nutrients / nitrogen (N)
#> 39                 ellenberg_salt                       Ellenberg salt (S)
#> 40                     eive_light                           EIVE light (L)
#> 41               eive_temperature                     EIVE temperature (T)
#> 42                  eive_moisture                        EIVE moisture (M)
#> 43                  eive_reaction                        EIVE reaction (R)
#> 44                 eive_nutrients                       EIVE nutrients (N)
#> 45                      woodiness                                Woodiness
#> 46         photosynthetic_pathway                   Photosynthetic pathway
#> 47                    growth_form                              Growth form
#> 48                      life_form                      Raunkiaer life form
#> 49             dispersal_syndrome                       Dispersal syndrome
#> 50             pollination_vector                       Pollination vector
#> 51                   life_history                             Life history
#> 52                  flower_colour                            Flower colour
#> 53                     fruit_type                               Fruit type
#> 54                     diet_guild                               Diet guild
#> 55                  activity_time                       Diel activity time
#> 56                  foraging_mode                            Foraging mode
#> 57                      migration               Migratory behaviour (bird)
#> 58                     flightless                    Flightlessness (bird)
#> 59                       venomous                       Venomous (reptile)
#> 60                      sociality                          Sociality (bee)
#> 61                          lecty          Pollen host breadth (bee lecty)
#> 62               nesting_strategy                   Nesting strategy (bee)
#> 63                 territoriality                 Territoriality (odonate)
#> 64                  mate_guarding                  Mate guarding (odonate)
#> 65                    flight_mode                    Flight mode (odonate)
#> 66                     gram_stain                  Gram stain (prokaryote)
#> 67              oxygen_metabolism           Oxygen metabolism (prokaryote)
#> 68                     cell_shape                  Cell shape (prokaryote)
#> 69                       motility                    Motility (prokaryote)
#> 70                    sporulation                 Sporulation (prokaryote)
#> 71            fungal_trophic_mode                      Fungal trophic mode
#> 72               larval_nutrition                   Larval nutrition (bee)
#> 73                  calcification                    Calcification (algae)
#> 74                    gamete_type                      Gamete type (algae)
#> 75               algal_life_cycle          Life-cycle ploidy phase (algae)
#> 76                algal_substrate             Attachment substrate (algae)
#> 77                   bioturbation Bioturbation mode (benthic invertebrate)
#> 78                   living_habit      Living habit (benthic invertebrate)
#> 79                  feeding_guild     Feeding guild (benthic invertebrate)
#> 80              skeletal_rigidity            Skeletal rigidity (octocoral)
#> 81             colony_growth_form           Colony growth form (octocoral)
#> 82                      depth_min                            Minimum depth
#> 83                      depth_max                            Maximum depth
#> 84                  elevation_min                        Minimum elevation
#> 85                  elevation_max                        Maximum elevation
#> 86                     home_range                               Home range
#> 87                     range_size                    Geographic range size
#> 88                habitat_breadth                          Habitat breadth
#> 89              generation_length                        Generation length
#> 90                    weaning_age                              Weaning age
#> 91                           ldmc                  Leaf dry matter content
#> 92                    seed_length                              Seed length
#> 93                  vulnerability                 Vulnerability to fishing
#> 94                            itd   Inter-tegular distance (bee body size)
#> 95                  tongue_length                      Tongue length (bee)
#> 96                   diet_breadth                             Diet breadth
#> 97                   aspect_ratio                  Caudal fin aspect ratio
#> 98                body_elongation                   Body elongation (fish)
#> 99          vertical_eye_position             Vertical eye position (fish)
#> 100             relative_eye_size                 Relative eye size (fish)
#> 101            oral_gape_position                Oral gape position (fish)
#> 102     relative_maxillary_length         Relative maxillary length (fish)
#> 103            body_lateral_shape                Body lateral shape (fish)
#> 104         pectoral_fin_position             Pectoral fin position (fish)
#> 105             pectoral_fin_size                 Pectoral fin size (fish)
#> 106    caudal_peduncle_throttling        Caudal peduncle throttling (fish)
#> 107                   head_length                              Head length
#> 108                    head_width                               Head width
#> 109               forelimb_length                          Forelimb length
#> 110               hindlimb_length                          Hindlimb length
#> 111                 elytra_length                            Elytra length
#> 112                antenna_length                           Antenna length
#> 113               pronotum_length                          Pronotum length
#> 114                   cell_length                              Cell length
#> 115                    cell_width                               Cell width
#> 116                cell_biovolume              Cell biovolume (microalgae)
#> 117                cell_thickness              Cell thickness (microalgae)
#> 118             cell_surface_area           Cell surface area (microalgae)
#> 119               colony_diameter          Colony maximum diameter (coral)
#> 120               corallite_width                  Corallite width (coral)
#> 121               carapace_length                 Carapace length (turtle)
#> 122              colour_lightness           Body colour lightness (beetle)
#> 123                     voltinism         Voltinism (generations per year)
#> 124    optimal_growth_temperature  Optimal growth temperature (prokaryote)
#> 125                   genome_size                 Genome size (prokaryote)
#> 126             optimal_growth_ph           Optimal growth pH (prokaryote)
#> 127                    gc_content           Genome GC content (prokaryote)
#> 128           conservation_status                     IUCN Red List status
#> 129                    body_shape                        Body shape (fish)
#> 130              caudal_fin_shape                  Caudal fin shape (fish)
#> 131                  feeding_mode                      Feeding mode (fish)
#> 132                mouth_position                    Mouth position (fish)
#> 133                 air_breathing                     Air breathing (fish)
#> 134                 parental_care                     Parental care (fish)
#> 135                 sexual_system                            Sexual system
#> 136             reproductive_mode               Reproductive (parity) mode
#> 137            lichen_growth_form             Lichen growth form (thallus)
#> 138                     substrate                                Substrate
#> 139                    photobiont                        Lichen photobiont
#> 140         reproductive_strategy Reproductive strategy (sexual / asexual)
#> 141                     leaf_type                                Leaf type
#> 142                 deciduousness                       Leaf deciduousness
#> 143                        marine                           Marine habitat
#> 144                    freshwater                       Freshwater habitat
#> 145                   coloniality                              Coloniality
#> 146                 wave_exposure                 Wave exposure preference
#> 147                 water_clarity                 Water clarity preference
#> 148                zooxanthellate         Zooxanthellate symbiosis (coral)
#> 149                   wing_length                              Wing length
#> 150                      wingspan                     Wingspan (butterfly)
#> 151                   beak_length                     Beak length (culmen)
#> 152               hand_wing_index                          Hand-wing index
#> 153             von_bertalanffy_k   Von Bertalanffy growth coefficient (K)
#> 154          von_bertalanffy_linf Von Bertalanffy asymptotic length (Linf)
#> 155                   thermal_max                      Upper thermal limit
#> 156                   thermal_min                      Lower thermal limit
#> 157                    beak_width                               Beak width
#> 158                    beak_depth                               Beak depth
#> 159                 tarsus_length                            Tarsus length
#> 160                   tail_length                              Tail length
#> 161                    secondary1                   First-secondary length
#> 162                kipps_distance                          Kipp's distance
#> 163                 root_diameter                            Root diameter
#> 164          specific_root_length                     Specific root length
#> 165            specific_root_area                       Specific root area
#> 166           root_tissue_density                      Root tissue density
#> 167            root_mass_fraction                       Root mass fraction
#> 168       root_dry_matter_content                  Root dry matter content
#> 169          root_n_concentration              Root nitrogen concentration
#> 170                 rooting_depth                            Rooting depth
#> 171 root_mycorrhizal_colonization            Root mycorrhizal colonization
#> 172     age_at_first_reproduction                Age at first reproduction
#> 173                 hunting_guild                   Hunting guild (spider)
#> 174                  web_building                    Web building (spider)
#> 175               bioluminescence            Bioluminescence (zooplankton)
#> 176       diel_vertical_migration    Diel vertical migration (zooplankton)
#> 177                nest_structure                   Nest structure (birds)
#> 178                     nest_site                        Nest site (birds)
#> 179               nest_attachment                  Nest attachment (birds)
#> 180              mycorrhizal_type                         Mycorrhizal type
#>            kind                        unit n_sources
#> 1       numeric                           m         5
#> 2       numeric                          mg         7
#> 3       numeric                      mm2/mg         4
#> 4       numeric                       g/cm3         5
#> 5       numeric                         mm2         3
#> 6       numeric                          mm         1
#> 7       numeric                          mm         1
#> 8       numeric                        mg/g         2
#> 9       numeric                        mg/g         2
#> 10      numeric                          mm         2
#> 11      numeric                      months         3
#> 12      numeric                          yr         3
#> 13      numeric                           g        14
#> 14      numeric                          yr         9
#> 15      numeric        trophic level (~1-5)         6
#> 16      numeric offspring per clutch/litter         8
#> 17      numeric                          yr         6
#> 18      numeric                          yr         3
#> 19      numeric                        days         4
#> 20      numeric                        days         2
#> 21      numeric                          mm        11
#> 22      numeric                           W         2
#> 23      numeric                           g         2
#> 24      numeric                    per year         6
#> 25      numeric                          yr         4
#> 26      numeric                       count         2
#> 27      numeric             individuals/km2         3
#> 28      numeric                           g         4
#> 29      numeric                           g         1
#> 30      numeric                          mm         3
#> 31      numeric                          mm         3
#> 32      numeric                month (1-12)         2
#> 33      numeric                month (1-12)         2
#> 34      numeric               1-9 (classic)         3
#> 35      numeric               1-9 (classic)         2
#> 36      numeric              1-12 (classic)         3
#> 37      numeric               1-9 (classic)         3
#> 38      numeric               1-9 (classic)         3
#> 39      numeric               0-9 (classic)         3
#> 40      numeric                 0-10 (EIVE)         1
#> 41      numeric                 0-10 (EIVE)         1
#> 42      numeric                 0-10 (EIVE)         1
#> 43      numeric                 0-10 (EIVE)         1
#> 44      numeric                 0-10 (EIVE)         1
#> 45  categorical                        <NA>         4
#> 46  categorical                        <NA>         3
#> 47  categorical                        <NA>         4
#> 48  categorical                        <NA>         3
#> 49  categorical                        <NA>         5
#> 50  categorical                        <NA>         4
#> 51  categorical                        <NA>         2
#> 52  categorical                        <NA>         3
#> 53  categorical                        <NA>         3
#> 54  categorical                        <NA>         7
#> 55  categorical                        <NA>         6
#> 56  categorical                        <NA>         2
#> 57  categorical                        <NA>         1
#> 58  categorical                        <NA>         1
#> 59  categorical                        <NA>         1
#> 60  categorical                        <NA>         1
#> 61  categorical                        <NA>         1
#> 62  categorical                        <NA>         1
#> 63  categorical                        <NA>         1
#> 64  categorical                        <NA>         1
#> 65  categorical                        <NA>         1
#> 66  categorical                        <NA>         2
#> 67  categorical                        <NA>         2
#> 68  categorical                        <NA>         2
#> 69  categorical                        <NA>         2
#> 70  categorical                        <NA>         1
#> 71  categorical                        <NA>         2
#> 72  categorical                        <NA>         1
#> 73  categorical                        <NA>         1
#> 74  categorical                        <NA>         1
#> 75  categorical                        <NA>         1
#> 76  categorical                        <NA>         1
#> 77  categorical                        <NA>         2
#> 78  categorical                        <NA>         2
#> 79  categorical                        <NA>         2
#> 80  categorical                        <NA>         1
#> 81  categorical                        <NA>         1
#> 82      numeric                           m         6
#> 83      numeric                           m         6
#> 84      numeric                           m         3
#> 85      numeric                           m         3
#> 86      numeric                         km2         3
#> 87      numeric                         km2         3
#> 88      numeric                       count         4
#> 89      numeric                          yr         3
#> 90      numeric                        days         3
#> 91      numeric                        mg/g         4
#> 92      numeric                          mm         4
#> 93      numeric                 0-100 index         3
#> 94      numeric                          mm         2
#> 95      numeric                          mm         2
#> 96      numeric                       count         3
#> 97      numeric                       index         2
#> 98      numeric                       index         1
#> 99      numeric                       index         1
#> 100     numeric                       index         1
#> 101     numeric                       index         1
#> 102     numeric                       index         1
#> 103     numeric                       index         1
#> 104     numeric                       index         1
#> 105     numeric                       index         1
#> 106     numeric                       index         1
#> 107     numeric                          mm         2
#> 108     numeric                          mm         2
#> 109     numeric                          mm         1
#> 110     numeric                          mm         1
#> 111     numeric                          mm         1
#> 112     numeric                          mm         1
#> 113     numeric                          mm         1
#> 114     numeric                          um         2
#> 115     numeric                          um         2
#> 116     numeric                         um3         1
#> 117     numeric                          um         1
#> 118     numeric                         um2         1
#> 119     numeric                          cm         1
#> 120     numeric                          mm         1
#> 121     numeric                          mm         1
#> 122     numeric                       index         1
#> 123     numeric                    per year         2
#> 124     numeric                       deg C         2
#> 125     numeric                          bp         1
#> 126     numeric                          pH         2
#> 127     numeric                           %         1
#> 128 categorical                        <NA>         6
#> 129 categorical                        <NA>         3
#> 130 categorical                        <NA>         2
#> 131 categorical                        <NA>         1
#> 132 categorical                        <NA>         1
#> 133 categorical                        <NA>         1
#> 134 categorical                        <NA>         1
#> 135 categorical                        <NA>         4
#> 136 categorical                        <NA>         2
#> 137 categorical                        <NA>         1
#> 138 categorical                        <NA>         1
#> 139 categorical                        <NA>         1
#> 140 categorical                        <NA>         1
#> 141 categorical                        <NA>         2
#> 142 categorical                        <NA>         2
#> 143 categorical                        <NA>         3
#> 144 categorical                        <NA>         3
#> 145 categorical                        <NA>         2
#> 146 categorical                        <NA>         2
#> 147 categorical                        <NA>         2
#> 148 categorical                        <NA>         2
#> 149     numeric                          mm         2
#> 150     numeric                          mm         1
#> 151     numeric                          mm         1
#> 152     numeric                       index         1
#> 153     numeric                      per yr         2
#> 154     numeric                          cm         2
#> 155     numeric                       deg C         2
#> 156     numeric                       deg C         1
#> 157     numeric                          mm         1
#> 158     numeric                          mm         1
#> 159     numeric                          mm         1
#> 160     numeric                          mm         1
#> 161     numeric                          mm         1
#> 162     numeric                          mm         1
#> 163     numeric                          mm         2
#> 164     numeric                         m/g         2
#> 165     numeric                       cm2/g         1
#> 166     numeric                       g/cm3         1
#> 167     numeric                         g/g         2
#> 168     numeric                         g/g         2
#> 169     numeric                        mg/g         2
#> 170     numeric                           m         2
#> 171     numeric                     percent         1
#> 172     numeric                          yr         1
#> 173 categorical                        <NA>         1
#> 174 categorical                        <NA>         1
#> 175 categorical                        <NA>         1
#> 176 categorical                        <NA>         1
#> 177 categorical                        <NA>         1
#> 178 categorical                        <NA>         1
#> 179 categorical                        <NA>         1
#> 180 categorical                        <NA>         1
#>                                                                                                                                                  sources
#> 1                                                                                                                      gift, diaz, austraits, bien, brot
#> 2                                                                                                   diaz, gift, austraits, bien, brot, ecoflora, kew_sid
#> 3                                                                                                                                 leda, gift, bien, brot
#> 4                                                                                                                      gwdd, austraits, bien, gift, leda
#> 5                                                                                                                                  austraits, bien, brot
#> 6                                                                                                                                              austraits
#> 7                                                                                                                                              austraits
#> 8                                                                                                                                        austraits, bien
#> 9                                                                                                                                        austraits, bien
#> 10                                                                                                                                            bien, gift
#> 11                                                                                                                                 austraits, brot, bien
#> 12                                                                                                                                 bien, gift, austraits
#> 13  combine, amniote, pantheria, elton_traits, avonet, anage, phylacine, repttraits, fishbase, sealifebase, frugivoria, pottier, animaltraits, homerange
#> 14                                                                  anage, amniote, combine, pantheria, repttraits, chelonians, amphibio, beukhof, sheld
#> 15                                                                                      fishbase, beukhof, quimbayo, pelagic, arctic_traits, sealifebase
#> 16                                                                        amniote, combine, pantheria, anage, repttraits, amphibio, chelonians, birdbase
#> 17                                                                                                  anage, amniote, amphibio, chelonians, beukhof, sheld
#> 18                                                                                                                               anage, amniote, combine
#> 19                                                                                                                    anage, combine, pantheria, amniote
#> 20                                                                                                                                   amniote, chelonians
#> 21                                 combine, amniote, repttraits, amphibio, fishbase, sealifebase, huang_amph, pottier, spider_traits, zooplankton, sheld
#> 22                                                                                                                                   anage, animaltraits
#> 23                                                                                                                                 combine, animaltraits
#> 24                                                                                            amniote, combine, anage, pantheria, repttraits, chelonians
#> 25                                                                                                                    pantheria, combine, amniote, anage
#> 26                                                                                                                                    pantheria, combine
#> 27                                                                                                                      pantheria, combine, tetradensity
#> 28                                                                                                                    amniote, combine, pantheria, anage
#> 29                                                                                                                                               amniote
#> 30                                                                                                                       amniote, repttraits, chelonians
#> 31                                                                                                                       amniote, repttraits, chelonians
#> 32                                                                                                                                    baseflor, ecoflora
#> 33                                                                                                                                    baseflor, ecoflora
#> 34                                                                                                                               floraweb, ecoflora, bet
#> 35                                                                                                                                         floraweb, bet
#> 36                                                                                                                               floraweb, ecoflora, bet
#> 37                                                                                                                               floraweb, ecoflora, bet
#> 38                                                                                                                               floraweb, ecoflora, bet
#> 39                                                                                                                          floraweb, ecoflora, baseflor
#> 40                                                                                                                                                  eive
#> 41                                                                                                                                                  eive
#> 42                                                                                                                                                  eive
#> 43                                                                                                                                                  eive
#> 44                                                                                                                                                  eive
#> 45                                                                                                                          zanne, gift, austraits, bien
#> 46                                                                                                                             gift, austraits, ecoflora
#> 47                                                                                                                           gift, austraits, bien, brot
#> 48                                                                                                                              gift, ecoflora, floraweb
#> 49                                                                                                                 gift, austraits, leda, baseflor, brot
#> 50                                                                                                               baseflor, ecoflora, floraweb, austraits
#> 51                                                                                                                                       gift, austraits
#> 52                                                                                                                                  gift, baseflor, bien
#> 53                                                                                                                             gift, baseflor, austraits
#> 54                                                                     avonet, elton_traits, repttraits, chelonians, blanchard, parravicini, zooplankton
#> 55                                                                                   repttraits, chelonians, quimbayo, combine, pantheria, spider_traits
#> 56                                                                                                                                repttraits, chelonians
#> 57                                                                                                                                                avonet
#> 58                                                                                                                                              birdbase
#> 59                                                                                                                                            repttraits
#> 60                                                                                                                                           eupolltrait
#> 61                                                                                                                                           eupolltrait
#> 62                                                                                                                                           eupolltrait
#> 63                                                                                                                                               odonata
#> 64                                                                                                                                               odonata
#> 65                                                                                                                                               odonata
#> 66                                                                                                                                        madin, bacdive
#> 67                                                                                                                                        madin, bacdive
#> 68                                                                                                                                        madin, bacdive
#> 69                                                                                                                                        madin, bacdive
#> 70                                                                                                                                                 madin
#> 71                                                                                                                               funguild, fungal_traits
#> 72                                                                                                                                           eupolltrait
#> 73                                                                                                                                          algae_traits
#> 74                                                                                                                                          algae_traits
#> 75                                                                                                                                          algae_traits
#> 76                                                                                                                                          algae_traits
#> 77                                                                                                                                   arctic_traits, nztd
#> 78                                                                                                                                   arctic_traits, nztd
#> 79                                                                                                                                   arctic_traits, nztd
#> 80                                                                                                                                             octocoral
#> 81                                                                                                                                             octocoral
#> 82                                                                                     fishbase, sealifebase, quimbayo, pelagic, coral_traits, octocoral
#> 83                                                                                     fishbase, sealifebase, quimbayo, pelagic, coral_traits, octocoral
#> 84                                                                                                                       birdbase, repttraits, globtherm
#> 85                                                                                                                       birdbase, repttraits, globtherm
#> 86                                                                                                                         combine, homerange, pantheria
#> 87                                                                                                                      avonet, chelonians, coral_traits
#> 88                                                                                                              birdbase, combine, frugivoria, pantheria
#> 89                                                                                                                              bet, frugivoria, combine
#> 90                                                                                                                             amniote, anage, pantheria
#> 91                                                                                                                         leda, gift, brot, diaz_traits
#> 92                                                                                                                           austraits, leda, gift, bien
#> 93                                                                                                                       fishbase, sealifebase, quimbayo
#> 94                                                                                                                              bee_ostwald, eupolltrait
#> 95                                                                                                                              bee_ostwald, eupolltrait
#> 96                                                                                                                          combine, pantheria, birdbase
#> 97                                                                                                                                     beukhof, quimbayo
#> 98                                                                                                                                             fishmorph
#> 99                                                                                                                                             fishmorph
#> 100                                                                                                                                            fishmorph
#> 101                                                                                                                                            fishmorph
#> 102                                                                                                                                            fishmorph
#> 103                                                                                                                                            fishmorph
#> 104                                                                                                                                            fishmorph
#> 105                                                                                                                                            fishmorph
#> 106                                                                                                                                            fishmorph
#> 107                                                                                                                               huang_amph, saproxylic
#> 108                                                                                                                               huang_amph, saproxylic
#> 109                                                                                                                                           huang_amph
#> 110                                                                                                                                           huang_amph
#> 111                                                                                                                                           saproxylic
#> 112                                                                                                                                           saproxylic
#> 113                                                                                                                                           saproxylic
#> 114                                                                                                                                 rimet_phyto, bacdive
#> 115                                                                                                                                 rimet_phyto, bacdive
#> 116                                                                                                                                          rimet_phyto
#> 117                                                                                                                                          rimet_phyto
#> 118                                                                                                                                          rimet_phyto
#> 119                                                                                                                                         coral_traits
#> 120                                                                                                                                         coral_traits
#> 121                                                                                                                                           chelonians
#> 122                                                                                                                                           saproxylic
#> 123                                                                                                                        arthropod_traits, eupolltrait
#> 124                                                                                                                                       madin, bacdive
#> 125                                                                                                                                                madin
#> 126                                                                                                                                       madin, bacdive
#> 127                                                                                                                                                madin
#> 128                                                                                 conservation_status, birdbase, phylacine, quimbayo, pelagic, pottier
#> 129                                                                                                                           beukhof, quimbayo, pelagic
#> 130                                                                                                                                    beukhof, quimbayo
#> 131                                                                                                                                              beukhof
#> 132                                                                                                                                             quimbayo
#> 133                                                                                                                                             fishbase
#> 134                                                                                                                                              beukhof
#> 135                                                                                                          tree_of_sex, coral_traits, octocoral, sheld
#> 136                                                                                                                              repttraits, sharkipedia
#> 137                                                                                                                                               italic
#> 138                                                                                                                                               italic
#> 139                                                                                                                                               italic
#> 140                                                                                                                                               italic
#> 141                                                                                                                               austraits, diaz_traits
#> 142                                                                                                                                      gift, austraits
#> 143                                                                                                                      sealifebase, combine, phylacine
#> 144                                                                                                                      sealifebase, combine, phylacine
#> 145                                                                                                                              coral_traits, octocoral
#> 146                                                                                                                              coral_traits, octocoral
#> 147                                                                                                                              coral_traits, octocoral
#> 148                                                                                                                              coral_traits, octocoral
#> 149                                                                                                                                   avonet, saproxylic
#> 150                                                                                                                                            leptraits
#> 151                                                                                                                                               avonet
#> 152                                                                                                                                               avonet
#> 153                                                                                                                                 beukhof, sharkipedia
#> 154                                                                                                                                 beukhof, sharkipedia
#> 155                                                                                                                                   globtherm, pottier
#> 156                                                                                                                                            globtherm
#> 157                                                                                                                                               avonet
#> 158                                                                                                                                               avonet
#> 159                                                                                                                                               avonet
#> 160                                                                                                                                               avonet
#> 161                                                                                                                                               avonet
#> 162                                                                                                                                               avonet
#> 163                                                                                                                                     groot, austraits
#> 164                                                                                                                                     groot, austraits
#> 165                                                                                                                                                groot
#> 166                                                                                                                                                groot
#> 167                                                                                                                                     groot, austraits
#> 168                                                                                                                                     groot, austraits
#> 169                                                                                                                                     groot, austraits
#> 170                                                                                                                                          groot, brot
#> 171                                                                                                                                                groot
#> 172                                                                                                                                              combine
#> 173                                                                                                                                        spider_traits
#> 174                                                                                                                                        spider_traits
#> 175                                                                                                                                          zooplankton
#> 176                                                                                                                                          zooplankton
#> 177                                                                                                                                            nesttrait
#> 178                                                                                                                                            nesttrait
#> 179                                                                                                                                            nesttrait
#> 180                                                                                                                                           fungalroot